Cope with extra values from parse-time-string
Thanks to @ury-marshak for reporting this potential problem:

On 2019-08-05 parse-time-string in emacs git HEAD started returning an
additional tenth value (subseconds) which broke memento-mori.

On 2019-08-19 it seems they've reverted the change or rather made it
still return 9 values to be compatible with all the code out there.

Let's ensure we're compatible with any such future change anyway.

This commit is an alternative implementation of PR #9.